move wds constants to ee
