+++ app/client/src/modules/ui-builder/ui/wds/WDSModalWidget/widget/index.tsx
@@ -8,7 +8,7 @@ import React from "react";
 import { EventType } from "constants/AppsmithActionConstants/ActionConstants";
 import { SelectionRequestType } from "sagas/WidgetSelectUtils";
 import { ModalBody } from "@appsmith/wds";
-import { WDS_MODAL_WIDGET_CLASSNAME } from "modules/ui-builder/ui/wds/constants";
+import { WDS_MODAL_WIDGET_CLASSNAME } from "ee/modules/ui-builder/ui/wds/constants";
 import type { CanvasWidgetsReduxState } from "reducers/entityReducers/canvasWidgetsReducer";
 import type {
   CopiedWidgetData,
@@ -21,7 +21,7 @@ import { ModalLayoutProvider } from "layoutSystems/anvil/layoutComponents/ModalL
 import styles from "./styles.module.css";
 import { getAnvilWidgetDOMId } from "layoutSystems/common/utils/LayoutElementPositionsObserver/utils";
 import { widgetTypeClassname } from "widgets/WidgetUtils";
-import { AnvilDataAttributes } from "modules/ui-builder/ui/wds/constants";
+import { AnvilDataAttributes } from "ee/modules/ui-builder/ui/wds/constants";
 
 class WDSModalWidget extends BaseWidget<ModalWidgetProps, WidgetState> {
   static type = "WDS_MODAL_WIDGET";
