+++ app/client/src/modules/ui-builder/ui/wds/WDSZoneWidget/widget/index.tsx
@@ -24,7 +24,10 @@ import type { LayoutProps } from "layoutSystems/anvil/utils/anvilTypes";
 import type { ContainerWidgetProps } from "widgets/ContainerWidget/widget";
 import { ContainerComponent } from "modules/ui-builder/ui/wds/Container";
 import { LayoutProvider } from "layoutSystems/anvil/layoutComponents/LayoutProvider";
-import { Elevations, anvilWidgets } from "modules/ui-builder/ui/wds/constants";
+import {
+  Elevations,
+  anvilWidgets,
+} from "ee/modules/ui-builder/ui/wds/constants";
 import type { CanvasWidgetsReduxState } from "reducers/entityReducers/canvasWidgetsReducer";
 import type {
   CopiedWidgetData,
